adds support to skip classes in model diagrams.
This should be a problem for controllers, but for models sometimes there are dynamically generated model associations and there is no way to exclude via a file name. The papertrail gem and it's associated versions table and PaperTrail::Version model being a great example. It is on nearly every model we have so it really clutters up model diagrams and excluding it wasn't currently possible.